+++ docs/examples/metrics/prometheus.py
@@ -28,7 +28,7 @@ from opentelemetry.sdk.metrics.export.controller import PushController
 start_http_server(port=8000, addr="localhost")
 
 # Meter is responsible for creating and recording metrics
-metrics.set_preferred_meter_provider_implementation(lambda _: MeterProvider())
+metrics.set_meter_provider(MeterProvider())
 meter = metrics.get_meter(__name__)
 # exporter to export metrics to Prometheus
 prefix = "MyAppPrefix"