+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/__init__.py
@@ -13,12 +13,12 @@
 # limitations under the License.
 
 import logging
-from collections import OrderedDict
+import threading
 from typing import Dict, Sequence, Tuple, Type
 
 from opentelemetry import metrics as metrics_api
 from opentelemetry.sdk.metrics.export.aggregate import Aggregator
-from opentelemetry.sdk.metrics.export.batcher import Batcher, UngroupedBatcher
+from opentelemetry.sdk.metrics.export.batcher import UngroupedBatcher
 from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.util.instrumentation import InstrumentationInfo
 from opentelemetry.util import time_ns
@@ -48,17 +48,18 @@ class LabelSet(metrics_api.LabelSet):
         return self._encoded == other._encoded
 
 
-class BaseHandle:
-    """The base handle class containing common behavior for all handles.
+class BaseBoundInstrument:
+    """Class containing common behavior for all bound metric instruments.
 
-    Handles are responsible for operating on data for metric instruments for a
-    specific set of labels.
+    Bound metric instruments are responsible for operating on data for metric
+    instruments for a specific set of labels.
 
     Args:
-        value_type: The type of values this handle holds (int, float).
+        value_type: The type of values for this bound instrument (int, float).
         enabled: True if the originating instrument is enabled.
-        aggregator: The aggregator for this handle. Will handle aggregation
-            upon updates and checkpointing of values for exporting.
+        aggregator: The aggregator for this bound metric instrument. Will
+            handle aggregation upon updates and checkpointing of values for
+            exporting.
     """
 
     def __init__(
@@ -71,6 +72,8 @@ class BaseHandle:
         self.enabled = enabled
         self.aggregator = aggregator
         self.last_update_timestamp = time_ns()
+        self._ref_count = 0
+        self._ref_count_lock = threading.Lock()
 
     def _validate_update(self, value: metrics_api.ValueT) -> bool:
         if not self.enabled:
@@ -86,6 +89,21 @@ class BaseHandle:
         self.last_update_timestamp = time_ns()
         self.aggregator.update(value)
 
+    def release(self):
+        self.decrease_ref_count()
+
+    def decrease_ref_count(self):
+        with self._ref_count_lock:
+            self._ref_count -= 1
+
+    def increase_ref_count(self):
+        with self._ref_count_lock:
+            self._ref_count += 1
+
+    def ref_count(self):
+        with self._ref_count_lock:
+            return self._ref_count
+
     def __repr__(self):
         return '{}(data="{}", last_update_timestamp={})'.format(
             type(self).__name__,
@@ -94,16 +112,16 @@ class BaseHandle:
         )
 
 
-class CounterHandle(metrics_api.CounterHandle, BaseHandle):
+class BoundCounter(metrics_api.BoundCounter, BaseBoundInstrument):
     def add(self, value: metrics_api.ValueT) -> None:
-        """See `opentelemetry.metrics.CounterHandle.add`."""
+        """See `opentelemetry.metrics.BoundCounter.add`."""
         if self._validate_update(value):
             self.update(value)
 
 
-class MeasureHandle(metrics_api.MeasureHandle, BaseHandle):
+class BoundMeasure(metrics_api.BoundMeasure, BaseBoundInstrument):
     def record(self, value: metrics_api.ValueT) -> None:
-        """See `opentelemetry.metrics.MeasureHandle.record`."""
+        """See `opentelemetry.metrics.BoundMeasure.record`."""
         if self._validate_update(value):
             self.update(value)
 
@@ -113,11 +131,11 @@ class Metric(metrics_api.Metric):
 
     Also known as metric instrument. This is the class that is used to
     represent a metric that is to be continuously recorded and tracked. Each
-    metric has a set of handles that are created from the metric. See
-    `BaseHandle` for information on handles.
+    metric has a set of bound metrics that are created from the metric. See
+    `BaseBoundInstrument` for information on bound metric instruments.
     """
 
-    HANDLE_TYPE = BaseHandle
+    BOUND_INSTR_TYPE = BaseBoundInstrument
 
     def __init__(
         self,
@@ -136,20 +154,23 @@ class Metric(metrics_api.Metric):
         self.meter = meter
         self.label_keys = label_keys
         self.enabled = enabled
-        self.handles = {}
+        self.bound_instruments = {}
+        self.bound_instruments_lock = threading.Lock()
 
-    def get_handle(self, label_set: LabelSet) -> BaseHandle:
-        """See `opentelemetry.metrics.Metric.get_handle`."""
-        handle = self.handles.get(label_set)
-        if not handle:
-            handle = self.HANDLE_TYPE(
-                self.value_type,
-                self.enabled,
-                # Aggregator will be created based off type of metric
-                self.meter.batcher.aggregator_for(self.__class__),
-            )
-            self.handles[label_set] = handle
-        return handle
+    def bind(self, label_set: LabelSet) -> BaseBoundInstrument:
+        """See `opentelemetry.metrics.Metric.bind`."""
+        with self.bound_instruments_lock:
+            bound_instrument = self.bound_instruments.get(label_set)
+            if bound_instrument is None:
+                bound_instrument = self.BOUND_INSTR_TYPE(
+                    self.value_type,
+                    self.enabled,
+                    # Aggregator will be created based off type of metric
+                    self.meter.batcher.aggregator_for(self.__class__),
+                )
+                self.bound_instruments[label_set] = bound_instrument
+        bound_instrument.increase_ref_count()
+        return bound_instrument
 
     def __repr__(self):
         return '{}(name="{}", description="{}")'.format(
@@ -163,31 +184,13 @@ class Counter(Metric, metrics_api.Counter):
     """See `opentelemetry.metrics.Counter`.
     """
 
-    HANDLE_TYPE = CounterHandle
-
-    def __init__(
-        self,
-        name: str,
-        description: str,
-        unit: str,
-        value_type: Type[metrics_api.ValueT],
-        meter: "Meter",
-        label_keys: Sequence[str] = (),
-        enabled: bool = True,
-    ):
-        super().__init__(
-            name,
-            description,
-            unit,
-            value_type,
-            meter,
-            label_keys=label_keys,
-            enabled=enabled,
-        )
+    BOUND_INSTR_TYPE = BoundCounter
 
     def add(self, value: metrics_api.ValueT, label_set: LabelSet) -> None:
         """See `opentelemetry.metrics.Counter.add`."""
-        self.get_handle(label_set).add(value)
+        bound_intrument = self.bind(label_set)
+        bound_intrument.add(value)
+        bound_intrument.release()
 
     UPDATE_FUNCTION = add
 
@@ -195,11 +198,13 @@ class Counter(Metric, metrics_api.Counter):
 class Measure(Metric, metrics_api.Measure):
     """See `opentelemetry.metrics.Measure`."""
 
-    HANDLE_TYPE = MeasureHandle
+    BOUND_INSTR_TYPE = BoundMeasure
 
     def record(self, value: metrics_api.ValueT, label_set: LabelSet) -> None:
         """See `opentelemetry.metrics.Measure.record`."""
-        self.get_handle(label_set).record(value)
+        bound_intrument = self.bind(label_set)
+        bound_intrument.record(value)
+        bound_intrument.release()
 
     UPDATE_FUNCTION = record
 
@@ -299,6 +304,7 @@ class Meter(metrics_api.Meter):
         self.metrics = set()
         self.observers = set()
         self.batcher = UngroupedBatcher(stateful)
+        self.observers_lock = threading.Lock()
         self.resource = resource
 
     def collect(self) -> None:
@@ -314,26 +320,39 @@ class Meter(metrics_api.Meter):
 
     def _collect_metrics(self) -> None:
         for metric in self.metrics:
-            if metric.enabled:
-                for label_set, handle in metric.handles.items():
+            if not metric.enabled:
+                continue
+
+            to_remove = []
+
+            with metric.bound_instruments_lock:
+                for label_set, bound_instr in metric.bound_instruments.items():
                     # TODO: Consider storing records in memory?
-                    record = Record(metric, label_set, handle.aggregator)
+                    record = Record(metric, label_set, bound_instr.aggregator)
                     # Checkpoints the current aggregators
                     # Applies different batching logic based on type of batcher
                     self.batcher.process(record)
 
+                    if bound_instr.ref_count() == 0:
+                        to_remove.append(label_set)
+
+                # Remove handles that were released
+                for label_set in to_remove:
+                    del metric.bound_instruments[label_set]
+
     def _collect_observers(self) -> None:
-        for observer in self.observers:
-            if not observer.enabled:
-                continue
+        with self.observers_lock:
+            for observer in self.observers:
+                if not observer.enabled:
+                    continue
 
-            # TODO: capture timestamp?
-            if not observer.run():
-                continue
+                # TODO: capture timestamp?
+                if not observer.run():
+                    continue
 
-            for label_set, aggregator in observer.aggregators.items():
-                record = Record(observer, label_set, aggregator)
-                self.batcher.process(record)
+                for label_set, aggregator in observer.aggregators.items():
+                    record = Record(observer, label_set, aggregator)
+                    self.batcher.process(record)
 
     def record_batch(
         self,
@@ -388,8 +407,13 @@ class Meter(metrics_api.Meter):
             label_keys,
             enabled,
         )
-        self.observers.add(ob)
+        with self.observers_lock:
+            self.observers.add(ob)
         return ob
+
+    def unregister_observer(self, observer: "Observer") -> None:
+        with self.observers_lock:
+            self.observers.remove(observer)
 
     def get_label_set(self, labels: Dict[str, str]):
         """See `opentelemetry.metrics.Meter.create_metric`.