+++ opentelemetry-sdk/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## Unreleased
 
+## 0.5b0
+
+Released 2020-03-16
+
+- Adding Correlation Context SDK and propagator
+  ([#471](https://github.com/open-telemetry/opentelemetry-python/pull/471))
+- Adding OT Collector metrics exporter
+  ([#454](https://github.com/open-telemetry/opentelemetry-python/pull/454))
+- Improve validation of attributes
+  ([#460](https://github.com/open-telemetry/opentelemetry-python/pull/460))
+- Moving resources to sdk
+  ([#464](https://github.com/open-telemetry/opentelemetry-python/pull/464))
+- Re-raise errors caught in opentelemetry.sdk.trace.Tracer.use_span()
+  ([#469](https://github.com/open-telemetry/opentelemetry-python/pull/469))
+- Implement observer instrument
+  ([#425](https://github.com/open-telemetry/opentelemetry-python/pull/425))
+
 ## 0.4a0 
 
 Released 2020-02-21